breaking change: rename discovery.file to local.file_match
This commit renames discovery.file to local.file_match so that:

* The purpose of the component is more clear (i.e., to find files on
  the local filesystem which match a given pattern).

* To remove a point of confusion where `discovery.file` was thought to
  be Prometheus' file SD.

* To allow a component for Prometheus' file SD to be named
  discovery.file, which otherwise would've collided with the now-named
  local.file_match.

A redirect from the old documentation page is explicitly not configured
because we anticipate creating a new page called discovery.file, which
will have a warning about the rename.

Related to #3607.
