+++ docs/sources/flow/reference/components/loki.echo.md
@@ -46,7 +46,7 @@ This example creates a pipeline that reads log files from `/var/log` and
 prints log lines to echo:
 
 ```river
-discovery.file "varlog" {
+local.file_match "varlog" {
   path_targets = [{
     __path__ = "/var/log/*log",
     job      = "varlog",
@@ -54,7 +54,7 @@ discovery.file "varlog" {
 }
 
 loki.source.file "logs" {
-  targets    = discovery.file.varlog.targets
+  targets    = local.file_match.varlog.targets
   forward_to = [loki.echo.example.receiver]
 }
